aws-application-signals-try/pyproject.toml
[project]
name = "aws-application-signals-try"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aws-xray-sdk>=2.14.0",
    "flask>=3.1.1",
    "opentelemetry-api>=1.34.0",
    "opentelemetry-exporter-otlp>=1.34.0",
    "opentelemetry-instrumentation-flask>=0.55b0",
    "opentelemetry-sdk>=1.34.0",
]
